+++ lib/g5_authenticatable_api/helpers/grape.rb
@@ -1,4 +1,5 @@
-require 'g5_authenticatable_api/token_validator'
+require 'g5_authenticatable_api/services/token_validator'
+require 'g5_authenticatable_api/services/user_fetcher'
 
 module G5AuthenticatableApi
   module Helpers
@@ -7,14 +8,37 @@ module G5AuthenticatableApi
         raise_auth_error if !token_validator.valid?
       end
 
+      def token_data
+        @token_data ||= token_info.token_data
+      end
+
+      def current_user
+        @current_user ||= user_fetcher.current_user
+      end
+
+      def access_token
+        @access_token ||= token_info.access_token
+      end
+
       def warden
         env['warden']
       end
 
       private
+      def request
+        Rack::Request.new(env)
+      end
+
+      def token_info
+        @token_info ||= Services::TokenInfo.new(request.params, headers, warden)
+      end
+
       def token_validator
-        request = Rack::Request.new(env)
-        @token_validator ||= TokenValidator.new(request.params, headers, warden)
+        @token_validator ||= Services::TokenValidator.new(request.params, headers, warden)
+      end
+
+      def user_fetcher
+        @user_fetcher ||= Services::UserFetcher.new(request.params, headers, warden)
       end
 
       def raise_auth_error